requirements.txt
numpy==1.21.3
pandas==1.3.3
matplotlib==3.5.0
seaborn==0.11.2
geopy==2.2.0
python-dotenv==0.19.2
pyfiglet==0.8.post1